Remove graphql codes
