+++ scripts/leaderboard/.env.sample
@@ -8,5 +8,4 @@ DATABASE_URL="postgresql://postgres:pass@localhost:5432/postgres"
 
 SUPABASE_URL=https://example.supabase.co
 SUPABASE_ANON_KEY=sample
-GRAPHQL_URI=http://example.com/graphql
