--- scripts/leaderboard/app/graphql/apollo-client.ts
@@ -1,14 +1,0 @@
-import { ApolloClient, InMemoryCache } from '@apollo/client';
-
-export const client = new ApolloClient({
-  uri: GRAPHQL_URI,
-  headers: {
-    apiKey: SUPABASE_ANON_KEY
-  },
-  cache: new InMemoryCache(),
-  defaultOptions: {
-    watchQuery: {
-      fetchPolicy: 'cache-and-network',
-    },
-  },
-});
