--- scripts/leaderboard/app/graphql/mutations/queue.graphql
@@ -1,14 +1,0 @@
-mutation lineup($teamId: UUID!, $pageUrl: String!) {
-  insertIntoQueueCollection(objects: [{teamId: $teamId, status: "RUNNING"}]) {
-    records {
-      id
-      teamId
-    }
-  }
-  updateTeamCollection(filter: { id: { eq: $teamId } }, set: { pageUrl: $pageUrl }) {
-    records {
-      id
-      pageUrl
-    }
-  }
-}
