--- scripts/leaderboard/app/graphql/mutations/signup.graphql
@@ -1,12 +1,0 @@
-mutation signup($email: String!, $name: String) {
-  insertIntoUserCollection(objects: [{
-    email: $email
-    name: $name
-  }]) {
-    records {
-      id
-      email
-      name
-    }
-  }
-}
