--- scripts/leaderboard/app/graphql/mutations/team.graphql
@@ -1,37 +1,0 @@
-mutation createTeam ($name: String!, $pageUrl: String) {
-  insertIntoTeamCollection(objects: [{
-    name: $name
-    pageUrl: $pageUrl
-  }]) {
-    records {
-      id
-    }
-  }
-}
-
-mutation joinTeam ($teamId: UUID!, $email: String!) {
-  updateUserCollection(filter: { email: { eq: $email } }, set: { teamId: $teamId }) {
-    records {
-      id
-      team {
-        id
-        name
-        userCollection {
-          edges {
-            node {
-              name
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-mutation leaveTeam ($email: String!) {
-  updateUserCollection(filter: { email: { eq: $email } }, set: { teamId: null }) {
-    records {
-      id
-    }
-  }
-}
