--- scripts/leaderboard/app/graphql/queries/sample.graphql
@@ -1,9 +1,0 @@
-query sample {
-  queueCollection {
-    edges {
-      node {
-        id
-      }
-    }
-  }
-}
