+++ scripts/leaderboard/global.d.ts
@@ -1,3 +1,2 @@
 declare const SUPABASE_URL: string;
-declare const GRAPHQL_URI: string;
 declare const SUPABASE_ANON_KEY: string;
