+++ scripts/leaderboard/package.json
@@ -13,15 +13,12 @@
     "dev:remix": "remix watch",
     "dev:miniflare": "cross-env NODE_ENV=development miniflare ./build/index.js --watch",
     "dev": "remix build && run-p dev:*",
-    "start": "cross-env NODE_ENV=production miniflare ./build/index.js",
-    "generate": "graphql-codegen --config codegen.yml",
-    "generate:fetch": "node --no-warnings scripts/fetchGraphQLSchema"
+    "start": "cross-env NODE_ENV=production miniflare ./build/index.js"
   },
   "prisma": {
     "seed": "ts-node prisma/seed.ts"
   },
   "dependencies": {
-    "@apollo/client": "^3.5.10",
     "@chakra-ui/icons": "^1.1.7",
     "@chakra-ui/react": "^1.8.7",
     "@emotion/cache": "^11.7.1",
@@ -36,7 +33,6 @@
     "@supabase/supabase-js": "^1.33.3",
     "cross-env": "^7.0.3",
     "framer-motion": "^6.2.8",
-    "graphql": "^16.3.0",
     "prisma": "^3.11.1",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
@@ -52,20 +48,12 @@
     "@cloudflare/workers-types": "^3.4.0",
     "@esbuild-plugins/node-globals-polyfill": "^0.1.1",
     "@faker-js/faker": "^6.1.2",
-    "@graphql-codegen/cli": "2.6.2",
-    "@graphql-codegen/fragment-matcher": "3.2.1",
-    "@graphql-codegen/introspection": "2.1.1",
-    "@graphql-codegen/typescript": "2.4.8",
-    "@graphql-codegen/typescript-graphql-files-modules": "2.1.1",
-    "@graphql-codegen/typescript-operations": "2.3.5",
-    "@graphql-codegen/typescript-react-apollo": "3.2.11",
     "@remix-run/dev": "^1.3.5",
     "@remix-run/eslint-config": "^1.3.5",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
     "esbuild-plugin-alias": "^0.2.1",
     "eslint": "^8.11.0",
-    "gradient-string": "^2.0.0",
     "miniflare": "^2.1.0",
     "npm-run-all": "^4.1.5",
     "remix-esbuild-override": "^3.0.2",
